Standardize admin UI role checks and harden users RLS policy
